requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.17,<4.0
psycopg2-binary>=2.9
pillow>=10.0
python-dotenv>=1.0
